Starting a music note recognition game
What is the note recognition game is a game designed to help the user learn and improve his ability to recognize musical notes. During the game, the notes will be played for the user and he will have to identify them by clicking on the corresponding note on a piano picture that is displayed. As the user progresses through the game, his progress in note recognition can be more consistent and accurate, so he can improve and progress in his musical abilities.
Existing Features:
Difficulty selection window (still working)
The game explanation window